requirements.txt
streamlit
Pillow
python-dotenv
google-generativeai
langchain
langchain-google-genai
ipykernel
